pyproject.toml
[project]
name = "youtube-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi>=0.128.0",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
    "youtube-transcript-api>=1.2.3",
]
